requirements.txt
fastapi
uvicorn[standard]
tefas-crawler

